Adjust tetra 28 to avoid rendering bug in vZome viewer
Online viewer version 98 doesn't handle the two consecutive identical SelectManifestation edits at the end of the original file. They should result in the blue strut being deselected as they do in desktop, but the bug causes it to remain selected in the online viewer. This commit simply removes the two superfluous edits altogether to avoid the bug.