Changed table selection UX
Added a highlight for the selected row.  Removed next/prev buttons
and the extra row at the top.  Added h1 index overlaying the viewer.

I tried "position: fixed" for the table heading, but then it covers the
first row and is not formatted to match the table columns.  More
CSS expertise is required to do what we want.